i clearly forgot all about commiting, i better get it together with commiting because i need them for the next project
okay so the triangles will be the death of me. i CAN not figure out how
to put them on my houses. i fixed it right side up but im in a slumpt. i
will continue to try and fix it until the due time lol. also idk what
happened but i tried to put a door and windows to my first house, and i
couldnt...? ill get back to that.